+++ cpp/cmake/templates/DOLFINXConfig.cmake.in
@@ -75,6 +75,10 @@ if(@SLEPC_FOUND@)
   endif()
 endif()
 
+if(@ADIOS2_FOUND@)
+  find_dependency(ADIOS2 2.8.1)
+endif()
+
 if(NOT TARGET dolfinx)
   include("${CMAKE_CURRENT_LIST_DIR}/DOLFINXTargets.cmake")
 endif()
