+++ cpp/dolfinx/CMakeLists.txt
@@ -112,7 +112,8 @@ endif()
 # ADIOS2
 if(ADIOS2_FOUND)
   target_compile_definitions(dolfinx PUBLIC HAS_ADIOS2)
-  target_link_libraries(dolfinx PRIVATE adios2::cxx11_mpi)
+  target_link_libraries(dolfinx PUBLIC adios2::cxx11_mpi)
+  # target_link_libraries(dolfinx PRIVATE adios2::cxx11_mpi)
 endif()
 
 # SLEPC
